Updated package to not conflict with original
- Include Unofficial in the package title
