+++ src/MediatR.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2013
 VisualStudioVersion = 12.0.31101.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MediatR", "MediatR\MediatR.csproj", "{D0CDC351-9F10-4531-BDFF-D9796487E9CE}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{8746A13B-A2BE-45A3-B261-782BC92689AA}"
 	ProjectSection(SolutionItems) = preProject
 		.nuget\NuGet.Config = .nuget\NuGet.Config
@@ -14,7 +12,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{8746A1
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Build", "Build", "{ED3AC9CD-B566-4242-A128-85BC11F91614}"
 	ProjectSection(SolutionItems) = preProject
-		MediatR\MediatR.nuspec = MediatR\MediatR.nuspec
+		Portable.MediatR\Portable.MediatR.nuspec = Portable.MediatR\Portable.MediatR.nuspec
 	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MediatR.Tests", "MediatR.Tests\MediatR.Tests.csproj", "{B81A30EB-BAC9-4FEA-8D65-FCBC2CDAC52B}"
@@ -25,16 +23,14 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MediatR.Examples", "MediatR
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MediatR.Examples.StructureMap", "MediatR.Examples.StructureMap\MediatR.Examples.StructureMap.csproj", "{C309CE69-7B72-4B6F-8456-A8936F740CF6}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Portable.MediatR", "Portable.MediatR\Portable.MediatR.csproj", "{D0CDC351-9F10-4531-BDFF-D9796487E9CE}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D0CDC351-9F10-4531-BDFF-D9796487E9CE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D0CDC351-9F10-4531-BDFF-D9796487E9CE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D0CDC351-9F10-4531-BDFF-D9796487E9CE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D0CDC351-9F10-4531-BDFF-D9796487E9CE}.Release|Any CPU.Build.0 = Release|Any CPU
 		{B81A30EB-BAC9-4FEA-8D65-FCBC2CDAC52B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{B81A30EB-BAC9-4FEA-8D65-FCBC2CDAC52B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B81A30EB-BAC9-4FEA-8D65-FCBC2CDAC52B}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -47,6 +43,10 @@ Global
 		{C309CE69-7B72-4B6F-8456-A8936F740CF6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C309CE69-7B72-4B6F-8456-A8936F740CF6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C309CE69-7B72-4B6F-8456-A8936F740CF6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D0CDC351-9F10-4531-BDFF-D9796487E9CE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D0CDC351-9F10-4531-BDFF-D9796487E9CE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D0CDC351-9F10-4531-BDFF-D9796487E9CE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D0CDC351-9F10-4531-BDFF-D9796487E9CE}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
