--- src/MediatR/MediatR.nuspec
@@ -1,15 +1,0 @@
-<?xml version="1.0"?>
-<package >
-  <metadata>
-    <id>$id$</id>
-    <version>$version$</version>
-    <title>$title$</title>
-    <authors>$author$</authors>
-    <owners>$author$</owners>
-    <licenseUrl>https://github.com/CollaboratingPlatypus/MediatR/MediatR/blob/master/LICENSE</licenseUrl>
-    <projectUrl>https://github.com/CollaboratingPlatypus/MediatR/MediatR</projectUrl>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>$description$</description>
-    <copyright>Copyright © Collaborating Platypus 2015</copyright>
-  </metadata>
-</package>
