--- src/MediatR/Properties/AssemblyInfo.cs
@@ -1,30 +1,0 @@
-﻿using System;
-using System.Reflection;
-using System.Resources;
-
-// General Information about an assembly is controlled through the following 
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-
-[assembly: AssemblyTitle("MediatR")]
-[assembly: AssemblyDescription("Simple mediator implementation based on Common Service Locator (with PCL version)")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("Collaborating Platypus")]
-[assembly: AssemblyProduct("MediatR")]
-[assembly: AssemblyCopyright("Copyright © Collaborating Platypus 2015")]
-[assembly: NeutralResourcesLanguage("en")]
-
-// Version information for an assembly consists of the following four values:
-//
-//      Major Version
-//      Minor Version 
-//      Build Number
-//      Revision
-//
-// You can specify all the values or you can default the Build and Revision Numbers 
-// by using the '*' as shown below:
-// [assembly: AssemblyVersion("1.0.*")]
-
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
-[assembly: CLSCompliant(true)]
