+++ src/Portable.MediatR/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Portable.CommonServiceLocator" version="1.2.2" targetFramework="portable-net45+win+wp80" />
+</packages>
